shader cleanup
